backend/vectorstore/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
faiss-gpu-cu12==1.11.0
sentence-transformers==5.0.0
beautifulsoup4==4.13.4
pandas==2.3.1
numpy<2.0.0
requests==2.32.4
httpx==0.28.1
pydantic==2.11.7
prometheus-client==0.22.1